ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
PyMuPDF==1.23.8
sentence-transformers==2.2.2
numpy==1.26.2
scikit-learn==1.3.2
python-dotenv==1.0.0
httpx==0.25.2
spacy==3.7.2
openai==1.3.0